code fix

+++ TCKit/TCFoundation/BaseUtlity/NSString+TCCypher.h
@@ -42,7 +42,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (unsigned long)CRC32B;
 - (nullable NSString *)CRC32BString;
-- (uint32_t)CRC32;
+- (unsigned long)CRC32;
 - (nullable NSString *)CRC32String;
 - (unsigned long)adler32;
 - (nullable NSString *)adler32String;
